Add RaftError enum & StorageError types.
Add the RaftError enum. It impls `Fail`. This provides a bit more data
to the users of this actor to understand the types of errors which may
be returned from the various interfaces of this system.

Add the StorageError type. It wraps a `dyn Fail` type, wich allows for
maximum flexibility for the underlying storage layer implementations.
The Raft actor is configured to stop when a storage error is
encountered, so the heap allocation should impose an absolutely
negligible cost.

Simplified the RaftStorage interface a bit by combining the
GetInitialState handler with the GetLastApplied handler. That latter has
been removed.