Finished up the election timeout system.
README updates on AdminCommand structure for controlling the Raft node
from outside of the Raft protocol itself.

A few new node states have been introduced for more clearly handling the
Raft node startup lifecycle. This provides more visibility into the
state of the system for the app which is using this system.

A bit of refactoring to remove unnecessary copy ops.