Major refactor of replication streams.
Refactored replication stream actor for more linear state based model.
This allows for a much more clear logical workflow, and does not involve
cancelling various tasks based on state changes. Instead, state changes
happen predictably, only in response to very specific events.

This update allows for control flow to proceed in exactly one of three
different routes depending on the state of the actor. Transitions
between states only happen as the current task reaches its end, and will
then cause the next state to be driven forward.

If there is nothing for the actor to do, the system will remain idle
with only the heartbeat loop running in the background. Only the receipt
of new entries to append to the target will cause the work to startup
again after an idle period.