Client request processing first take ... done.
This means that the core pieces of the Raft protocol are done. Still a
few pieces to implement and pin down.

Planning on doing a HUGE refactor of the Raft actor. There are a few
contending plans right now ... but they all revolve around implementing
logic more closely based on the state variants.