Client requests now represent single entry.
Batching is still possible, but it is represented as a single entry.

Updated the way client requests are passed through the system. Once a
payload receives an index assignment, it is upgraded and an Arc'd
version of the entry is retained for being given to the storage
interface and the replication streams.

Client request forwarding has been updated to stick a bit closer to the
Raft spec. If the Raft node is no longer the leader, it will return an
error with the original payload and the ID of the Raft node. This gives
applications more flexibility in terms of how they want to handle such
conditions instead of forcing applications to choose exactly one path.