Stabilization of tests after actix cancel_future workaround.
Replication streams have been refactored slightly to help with mental
load and readability.

ReplicationStream::drive_state_lagging has been updated to account for
conditions where the target has been in lagging state to the point where
it has fallen far enough behind to warrant a snapshot.

Updated AppendEntries RPC handler to account for updates in a more clear
and granular fashion.

Snapshot streaming is working, still running into a a few small kinks,
will hammer them out soon.

FINALLY FOUND THE ELECTION STABILITY ISSUE. Looks like the future
cancelling system doesn't quite cancel the futures as timely as needed.
Had to update the election timeout system to use an epoch timestamp for
a more precise mechanism to determine if an election timeout is really
needed.

Hammered out a few small issues in the voting system. Mostly just for
clarity, but should prove to keep the voting system clear and stable.

Added a timing mechanism to the replication streams lagging state. If an
error comes up, the stream will delay for a short amount of time in
order to keep from overwhelming the system.