A set of changes for a 0.4.2 release.
Fixed an issue where the value for `current_leader` was not being set
to `None` when becoming a candidate. This isn't really a *bug* per se,
as no functionality depended on this value as far as Raft is concerned,
but it is an issue that impacts the metrics system. This value is now
being updated properly.

Made the `messages::ClientPayload::new_base` constructor `pub(crate)`
instead of `pub`, which is what the intention was originally, but I
was apparently tired `:)`.

Implemented [#25](https://github.com/railgun-rs/actix-raft/issues/25).
Implementing Display+Error for the admin error types.