Everything is now implemented directly on top of Tokio.
The entire code base has been immensely simplified. Async/await is
largely responsible for the simplification here, however tokio has also
greatly contributed to the simplification of the code achieved here.

The RaftCore and the ReplicationCore types are now implemented as async
loops which delegate their state specific behaviors to other types which
take a mutable reference to the respective cores and then return to the
respective core loops when transitioning to other states.

The snapshot interface has been greatly simplified as well, and is now
based on AsyncRead/AsyncWrite. This mechanism was previously one of the
more complex components of the system, and is now quite simple.

The period log compaction/snapshotting system is now implemented as
well. This implementation is asynchronous to the core Raft tasks, can be
aborted if it needs to be interrupted due to needing to stream in a new
snapshot from the leader, and is fully spec compliant.

The new pattern for committing client requests to the cluster, the
pattern used for committing a new leader's initial payload, and the
pattern for cases where the new payload will commit a previously
uncommitted config change; all of these patterns are greatly
simplified and far more maintainable.

The above also includes automatic stepdown of a leader when it is the
leader of a cluster from which it has been removed. This functionality
was in place in earlier version of this crate, but the pattern has been
greatly simplified.

Added the `raft-essentials.md`. This is an implementation reference,
and is far better than having to review the Raft spec looking for
details on exact semantics and nuances for implementation.

[ci skip]