Finished `memstore`, the in-memory RaftStorage implementation.
Updated the RaftStorage system to use an associated type which is bound
by AsyncRead + AsyncWrite + AsyncSeek (and a few other traits). This
allows the write path to easily use the type directly via static
dispatch, and then for read-only cases, we can use dynamic dispatch to
ensure the snapshot is not written to when it is only exposed for
reading.

Updates a few aspects of how snapshotting is handled in core.
Implementing memstore helped to identify a few of the issues which would
come up in live systems.