Primary suite of integration tests are back in order.
Remodeled the error handling mechanisms throughout this project. Now,
Raft no longer cares about the error types used in the network & storage
layers. As long as they are std::error::Error + Send + Sync + 'static.

The `RaftStorage` interface has been slightly updated to offer a more
clear pattern for returning client request responses from the storage
layer. These are obviously different from errors, so now the
`RaftStorage::append_entry_to_log` method has been updated to return
`Option<R>` as the `Ok` variant. If `Some` is returned, then the client
request will be immediately returned indicating some application
specific response. The request will not be replicated and will not exist
in the log. If `None` is returned, then everything continues to work as
normal, and the request will be replicated, and ultimately applied to
the state machine returning an application specific data response.

Updated a few RaftStorage methods.
- `apply_entry_to_state_machine` now takes the index & the actual app
specific data payload to be applied. Previously, internal Raft entry
types were being passed into this method, and that was unnecessary. This
helps to keep things more nicely encapsulated, and helps to allow
storage engine implementors to focus on the core of their applications.
- `replicate_to_state_machine` now takes a dynamic dispatch iterator
over the index and application specific data payload of entries which
need to be applied to the state machine. This is similar to the above
change, and the intention is to help clarify the intent of this
interface.

Added a new RaftStorage method for getting the latest membership config
from the log. This is used in two locations now to ensure that the
correct membership is being used in cases where it may have been deleted
or overwritten.
- in the AppendEntries RPC handler, during conflict resolution where log
entries may need to be deleted.
- in the InstallSnapshot RPC handle, after snapshot installation is
finished.

Add tests for the stepdown system.