Updates to the log compaction interface.
The log index provided to the log compaction interface was a bit
misleading. When performing log compaction, the compaction can only
cover the breadth of the log up to the last applied log (obvs) and under
write load, this value may change quickly. As such, the expectations of
the log compaction interface have been refined and clarified.

Now, the only expectation is that the storage implementation will
export/checkpoint/snapshot its state machine, and then use the value of
that export's last applied log as the metadata indicating the breadth of
the log covered by the snapshot.