fix #112 : when a follower is removed, leader should stops sending log to it.
A leader adds all follower replication states to a hashset `nodes`, when
the leader is established.
But the leader does not do it when membership changed.
Thus when a follower is removed, the leader can not stop replication to
it because the follower is not in `nodes`.

The solution is to move replication state from `non_voters` to `nodes`.
So that next time a follower is removed the leader is able to remove the
replication from `nodes`.