feature: add metrics about leader
In LeaderState it also report metrics about the replication to other node when report metrics.

When switched to other state, LeaderState will be destroyed as long as
the cached replication metrics.

Other state report an `None` to raft core to override the previous
metrics data.

At some point the raft core, without knonwning the state, just report
metrics with an `Update::Ignore`, to indicate that leave replication
metrics intact.