fix: leader should not commit when there is no replication to voters.
When there is no replication to voters but there are replications to
non-voters, the leader did not check non-voters for a quorum but just
commits a log at once.

This cause the membership change log from a single node always commits.
E.g. start node 0, and non-voter 1, 2; then `change_membership({0, 1, 2})`,
It just commits the joint-log at once.
But according to raft paper, it should await a quorum of {0} and a
quorum of {0, 1, 2}.