change: to get last_log and membership, Storage should search for both logs and state machines.
Why:

depending on the impl, a RaftStore may have logs that are included in
the state machine still present.
This may be caused by a non-transactional impl of the store, e.g.
installing snapshot and removing logs are not atomic.

Thus when searching for last_log or last membership, a RaftStore should
search for both logs and state machine, and returns the greater one
that is found.

- Test: add test to prove these behaviors,
  which includes: `get_initial_state()` and `get_membership()`.

- Refactor: Make store tests a suite that could be applied to other
  impl.