change: remove PurgedMarker. keep logs clean
Changing log(add a PurgedMarker(original SnapshotPointer)) makes it
diffeicult to impl `install-snapshot` for a RaftStore without a lock
protecting both logs and state machine.

Adding a PurgedMarker and installing the snapshot has to be atomic in
storage layer. But usually logs and state machine are separated store.
e.g., logs are stored in fast flash disk and state machine is stored
some where else.

To get rid of the big lock, PurgedMarker is removed and installing a
snaphost does not need to keep consistent with logs any more.