change: introduce StorageError. RaftStorage gets rid of anyhow::Error
`StorageError` is an `enum` of DefensiveError and StorageIOError.
An error a RaftStorage impl returns could be a defensive check error
or an actual io operation error.

Why:

anyhow::Error is not enough to support the flow control in RaftCore.
It is typeless thus RaftCore can not decide what next to do
depending on the returned error.

Inside raft, anyhow::Error should never be used, although it could be used as
`source()` of some other error types.