change: `RaftStorage::finalize_snapshot_installation` is no more responsible to delete logs included in snapshot
A RaftStorage should be as simple and intuitive as possible.

One should be able to correctly impl a RaftStorage without reading the
guide but just by guessing what a trait method should do.

RaftCore is able to do the job of deleting logs that are included in
the state machine, RaftStorage should just do what is asked.