Fix: a Candidate should revert to Follower at once when a higher vote is seen
When a Candidate saw a higher vote, it store it at once.
Then no more further granted votes are valid to this candidate,
because vote they granted are changed.

Thus it was wrong to compare `last_log_id` before decide if to revert to
Follower.
The right way is to revert to Follower at once and stop the voting
procedure.

The test `leader_metrics()` sometimes fails because of this bug.