Feature: add ExampleClient to example-raft-kv
Add raft client in rust `ExampleClient` to show how to talk to a raft
cluster.
It implements 2 application APIs and several administrative APIs for managing or
monitor a raft cluster.

This client also shows how to deal `ForwardToLeader` error when a
**write** request is sent to a non-leader.

Also a new demo `test_cluster.rs` of how to set up a raft cluster with `ExampleClient` is
added.

Other modification:
- All API replies are made into a uniform form: `Result<T, E>`.
- `openraft` adds new error `NodeNotFound` and `Infallible`.

- fix: #158