Change: Add support for GUID-like `NodeId`
This adds support for 16-byte `NodeId` type, which can be used to store node GUIDs instead of simple `u64` which need additional mapping. In principle, it's also possible to store IPv6 address into the node ID, but currently it's meant to store GUIDs.

Ideal implementation would parameterize the Raft on `NodeId`, but that's a breaking change and fairly complex. This way, the user can activate the feature `guid_nodeid` to enable the GUID-based `NodeId`.

The implementation is not perfect also in another respect: the test crate actually using the GUID `NodeId` is not part of the regular `cargo test`. Not sure how to enable testing without polluting all other tests with GUID `NodeId`, failing their compilation. Thus, it has to be tested manually.