Change: RaftStorage: use `EffectiveMembership` instead of `Option<_>`
The value of membership config of an uninitialized raft-node should be:
```rust
EffectiveMembership {
    log_id: None,
    membership: Membership::default(),
}

```

I.e., there is always a membership config for either an initialized or
an uninitialized raft-node.
Because `EffectiveMembership` is not a value that could be `None`, but it is more like a container of values.

Changed struct:

```rust
struct InitialState {
    pub last_membership: EffectiveMembership<C>,
    ...
}
```

Changed methods:

```rust
RaftStorage::get_membership() -> Result<EffectiveMembership<C>, _>;
RaftStorage::last_applied_state() -> Result<(_, EffectiveMembership<C>), _>;
```