Refactor: when installing snapshot, delete logs included in snapshot
A local log that is included in snapshot may be inconsistent with the leader.
It has to purge all of them to prevent these log form being replicated, when this node becomes leader.

When deleting these logs, it does not need to consider the `max_applied_log_to_keep` config.
Thus `purge_applied_logs()` does not need to deal with the condition
that `last_log_id` is inconsistent with `last_applied`.