Refactor: add Progress to track replication state and to calculate committed log id
Progress is stored in a `Vec<(ID, VALUE)>`.
Calling `update(id, new_value)` when replication to target `id` is
updated to `new_value`.

Algorithm:

Only when the **previous value** is less or equal the committed,
and the **new value** is greater than the currently committed,
there is possibly an update to the committed.

This way it gets rid of a portion of unnecessary re-calculation of committed,
and avoids unnecessary sorting: progresses are kept in order and only values greater than committed need to
sort.

E.g., given 3 ids with value `1,3,5`,
the committed is `3` and assumes a majority quorum set is used.
Then:
- update(a, 6): nothing to do: committed is still 3;
- update(b, 4): re-calc:       committed becomes 4;
- update(b, 6): re-calc:       committed becomes 5;
- update(c, 2): nothing to do: committed is still 3;
- update(c, 3): nothing to do: committed is still 3;
- update(c, 4): re-calc:       committed becomes 4;
- update(c, 6): re-calc:       committed becomes 5;

```text
a -----------+-------->
b -------+------------>
c ---+---------------->
------------------------------
     1   3   5
```

A micro benchmark shows that `is_quorum()` is called 1/4 times of the
calls to `update()`:

```
Stat { update_count: 42997501, move_count: 10749381, is_quorum_count: 10749399 }
...::progress_update_01234_567   ... bench:           8 ns/iter (+/- 0)
```

- Fix: #396