Fix: avoid creating log-id with unitialized `matched.leader_id`.
When waiting for a newly added learner to become up to date,
it tries to compare last-log-id and the reported `matched` replication
state.
But the `matched` may have not yet receive any update and is
unitialized, in such case, it tries to create a temp LogId with
`leader_id(0, 0)`, which is illegal.

The fix is simple: do not use log-id. Just calculating replication lag by log index.

Add test to reproduce it: openraft/tests/membership/t99_issue_471_adding_learner_uses_uninit_leader_id.rs

- Fix: #471