Change: `ErrorSubject::Snapshot(SnapshotSignature)`
Change `ErrorSubject::Snapshot(SnapshotMeta)` to `ErrorSubject::Snapshot(SnapshotSignature)`.

`SnapshotSignature` is the same as `SnapshotMeta` except it does not include
`Membership` information.
This way errors do not have to depend on type `Node`, which is used in
`Membership` and it is a application specific type.

Then when a user-defined generic type `NodeData` is introduced, error
types do not need to change.

- Part of: #480