Change: `Membership.nodes` remove `Option` from value
Before this commit, the value of `Membership.nodes` is `Option<N:
Node>`: `Membership.nodes: BTreeMap<NID, Option<N>>`

The value does not have to be an `Option`.
If an application does not need openraft to store the `Node` data, it
can just implement `trait Node` with an empty struct, or just use
`BasicNode` as a placeholder.

- Using `Option<N>` as the value is a legacy and since #480 is merged, we
  do not need the `Option` any more.