Refactor: remove must_include:LogId from NeedSnapshot
When a replication task ask the RaftCore for a snapshot, it does not
need to specify what log id it lacks of any more.

Because any log a replication needs but does not exist, it is purged.
And a purged log id has to be included in a snapshot.
Thus the current snapshot must include to log the replication needs.