Fix: delay leader step down
When a membership that removes the leader is committed,
the leader continue to work for a short while before reverting to a learner.
This way, let the leader replicate the `membership-log-is-committed` message to followers.

Otherwise, if the leader step down at once, the follower might have to re-commit the membership log
again.

After committing the membership log that does not contain the leader,
the leader will step down in the next `tick`.