Refactor: specify the default value when building Progress
This is the first step to extend Progress to hold complete replication
state: the `matching` log id and the `max_possible_matching` log index.
I.e., progress should be `Progress<ID, (Option<LogId>, Option<LogIndex>), _>`.

Because `max_possible_matching` is decided by the leader, the default
value of a progress entry must be specified when it is created.