Feature: add config: send_snapshot_timeout
The timeouts for sending non-last and last snapshot segments may be
different: when a node receives the last segment, it installs the
snapshot thus it takes longer time.

Thus `send_snapshot_timeout` is for sending non-last segment.
The `install_snapshot_timeout` is for sending last segment, and is used
as default timeout for non-last segment if `send_snapshot_timeout` is 0.