Refactor: use the last snapshot when ReplicationCore asks
When a ReplicationCore asks RaftCore for a snapshot, RaftCore just
sends back the last built snapshot, it does not have to build a new one.
ReplciationCore guarantees it asks for a snapshot only when the last
snapshot fulfill its need: i.e., when it tries to replicate a purged
log(which must be included in the last snapshot).