Refactor: Remove repliation algorithm from ReplicationCore
Let Engine totally decide what to do, ReplicationCore becomes a passive executor.
Engine emit a replication request and let ReplicationCore execute it.
But it is not an exact one request one reply model, e.g., Engine does not expect the reply for a broadcast-committed request.

Engine keeps the track of every inflight replication request, and update
the `inflight` state when a reply is received from ReplicationCore.
A new type `Inflight` is introduced to manage the state of inflight
replication data, e.g., a series of log entries or a snapshot.

ReplicationCore does not need to ask for RaftCore for snapshot anymore.
But instead, if snapshot replication is needed, Engine will just send a
snapshot to ReplicationCore.

Introduce `ReplicationHandler` and `LogHandler` as a sub set function
container of Engine, in order to make `Engine` not that big.

Add more `Validate` implementations, to assert internal consistency.

Add ProgressEntry::next_send() to calculate what entries/snapshot to send

Remove ProgressEntry.mid; It can be calculated when needed.

- Fix: #650