Change: remove RaftNetworkFactory::ConnectionError and AddLearnerError::NetworkError
`RaftNetworkFactory::new_client()` does not return an error because
openraft can only ignore it.  Therefore it should **not** create a
connection but rather a client that will connect when required.  Thus
there is chance it will build a client that is unable to send out
anything, e.g., in case the Node network address is configured
incorrectly.

Because of the above change, And `AddLearnerError` will not include a
NetworkError any more, because when adding a learner, the connectivity
can not be effectively detected.

Upgrade tip:

Just update the application network implementation so that it compiles.