Change: remove AddLearnerResponse and AddLearnerError
In openraft adds a learner is done by committing a membership config
log, which is almost the same as committing any log.

`AddLearnerResponse` contains a field `matched` to indicate the
replication state to the learner, which is not included in
`ClientWriteResponse`. This information can be retrieved via
`Raft::metrics()`.

Therefore to keep the API simple, replace `AddLearnerResponse` with
`ClientWriteResponse`.

Behavior change: adding a learner always commit a new membership config
log, no matter if it already exists in membership.
To avoid duplicated add, an application should check existence first by
examining `Raft::metrics()`

- Fix: #679

Upgrade tips:

- Replace AddLearnerResponse with ClientWriteResponse
- Replace AddLearnerError with ClientWriteError

Passes the application compilation.

See the changes in examples/.