Change: remove non-blocking membership change
When changing membership in nonblocking mode, the leader submits a
membership config log but does not wait for the log to be committed.

This is useless because the caller has to assert the log is
committed, by periodically querying the metrics of a raft node, until it
is finally committed. Which actually makes it a blocking routine.

API changes:

- Removes `allow_lagging` paramenter from `Raft::change_membership()`
- Removes error `LearnerIsLagging`

Upgrade tip:

Adjust API calls to make it compile.

Refactor: move `leader_append_entries()` to `LeaderHandler`.