Change: trait IntoNodes adds two new method has_nodes() and node_ids()
`trait IntoNodes` converts types `T` such as `Vec` or `BTreeSet` into
`BTreeMap<NID, Node>`.

This patch changes the functionality of the `IntoNodes` trait to provide
two new methods `has_nodes()` and `node_ids()`, in addition to the existing
`into_nodes()` method. The `has_nodes()` method returns true if the type `T`
contains any `Node` objects, and `node_ids()` returns a `Vec` of the `NodeId`
objects associated with the `Node` objects in `T`.

Refactor:

The patch also refactors the `Membership::next_safe()` method to return an
`Err(LearnerNotFound)` if it attempts to build a `Membership` object
containing a `voter_id` that does not correspond to any `Node`.