Refactor: Adds a new method `ensure_last_membership_committed()` to the `MembershipState`
The function checks if the current effective membership configuration is
committed to the Raft log or not.

If it is committed, the function returns `Ok()`, otherwise, it returns an
error `Err(InProgress)`.

`create_updated_membership()` has been modified to use the new
`ensure_last_membership_committed()` to validate the committed and effective
configurations of the membership before building a new membership
configuration.
If the committed and effective membership configurations are not the
same, the `create_updated_membership()` function now returns a
`ChangeMembershipError::InProgress` error.

The patch also adds new tests to test the new
`ensure_last_membership_committed()` and modified `create_updated_membership()`
functions.