Fix: when adding a learner, ensure the last membership is committed
Previously, when adding a learner to a Raft cluster, the last membership was not
always marked as committed, which could cause issues when a follower tried to
truncate logs by reverting to the last committed membership. To prevent this
issue, we have updated the code to ensure the last membership is committed when
adding a learner.

In addition to this fix, we have also made several refactoring changes,
including refining method names for trait `Coherent`, renaming
`Membership::next_safe()` to `next_coherent()` for consistency, and updating
enum `ChangeMembers` to include more variants for adding and removing learners.
We have also removed `RaftCore::add_learner()` in favor of using
`change_membership()` for all membership operations, and added a `ChangeHandler`
to build new membership configurations for change-membership requests.

Finally, we have updated the `Membership` API with a new method `new_with_nodes()`
for building a new membership configuration, and moved the validation check
out into a separate function, `ensure_valid()`. Validation is now done only
when needed.