Change: move default implemented method from trait `RaftLogReader` to `StorageHelper`
Function `get_log_entries()` and `try_get_log_entry()` are provided by
trait `RaftLogReader` with default implementations. However, they do not
need to be part of this trait and an application does not have to
implement them.

Therefore in this patch they are moved to `StorageHelper` struct, which
provides additional storage access methods that are built based on the
`RaftStorage` trait.