Change: replace EffectiveMembership with StoredMembership in RaftStorage
`EffectiveMembership` is a struct used at runtime, which contains
additional information such as an optimized `QuorumSet` implementation
that has different structure from a `Membership`.

To better separate concerns, a new struct called `StoredMembership` has
been introduced specifically for storage purpose. It contains only the
information that needs to be stored in storage. Therefore,
`StoredMembership` is used instead of `EffectiveMembership` in
RaftStorage.

Upgrade tip:

Replace `EffectiveMembership` with `StoredMembership` in an application.

Fields in `EffectiveMembership` are made private and can be accessed via
corresponding methods such as: `EffectiveMembership.log_id` and
`EffectiveMembership.membership` should be replaced with
`EffectiveMembership::log_id()` and `EffectiveMembership::membership()`.