Improve: fix: delay election if a greater last log id is seen
If this node sees a greater last-log-id on another node, it will be less
likely to be elected as a leader.
In this case, it is necessary to sleep for a longer period of time
`smaller_log_timeout` so that other nodes with a greater last-log-id
have a chance to elect themselves.

Fix: such as state should be kept until next election, i.e., it should
be a field of `Engine` instead of a `field` of `internal_server_state`.
And this value should be greater than the `election_timeout` of every other node.