Refactor: add `input_entries` to RaftCore
`RaftCore.input_entries` stores input log entries to be appended.
These entries comes from user calls to `RaftCore::client_write` for a leader, or
`RaftCore::append_entries` for a follower or learner.

`input_entries` serves as the sole storage for all log entries that are pending to be
written to `RaftStorage`. Group committing will be done on this storage.